+++ fast-data-switch-admin/src/main/java/cn/zhumingwu/dataswitch/admin/route/strategy/ExecutorRouteLFU.java
@@ -75,7 +75,7 @@ public class ExecutorRouteLFU extends ExecutorRouter {
 
     @Override
     public Result<String> route(TriggerParam triggerParam, List<String> addressList) {
-        String address = route(triggerParam.getJobId(), addressList);
+        String address = route(triggerParam.getJobId().intValue(), addressList);
         return Result.success(address);
     }
 }
