+++ fast-data-switch-admin/src/main/java/cn/zhumingwu/dataswitch/admin/route/strategy/ExecutorRouteRound.java
@@ -12,11 +12,10 @@ import java.util.concurrent.ConcurrentMap;
 
 public class ExecutorRouteRound extends ExecutorRouter {
 
-    private static ConcurrentMap<Integer, Integer> routeCountEachJob =
-            new ConcurrentHashMap<Integer, Integer>();
+    private static ConcurrentMap<Long, Integer> routeCountEachJob = new ConcurrentHashMap<Long, Integer>();
     private static long CACHE_VALID_TIME = 0;
 
-    private static int count(int jobId) {
+    private static int count(Long jobId) {
         // cache clear
         if (System.currentTimeMillis() > CACHE_VALID_TIME) {
             routeCountEachJob.clear();
