+++ fast-data-switch-admin/src/main/java/cn/zhumingwu/dataswitch/admin/thread/JobTriggerPoolHelper.java
@@ -21,7 +21,7 @@ public class JobTriggerPoolHelper {
     private ThreadPoolExecutor slowTriggerPool = null;
     // job timeout count
     private volatile long minTim = System.currentTimeMillis() / 60000; // ms > min
-    private volatile ConcurrentMap<Integer, AtomicInteger> jobTimeoutCountMap = new ConcurrentHashMap<>();
+    private volatile ConcurrentMap<Long, AtomicInteger> jobTimeoutCountMap = new ConcurrentHashMap<>();
 
     public static void toStart() {
         helper.start();
@@ -41,7 +41,7 @@ public class JobTriggerPoolHelper {
      * @param executorParam         null: use job param not null: cover job param
      */
     public static void trigger(
-            int jobId,
+            Long jobId,
             TriggerTypeEnum triggerType,
             int failRetryCount,
             String executorShardingParam,
@@ -88,7 +88,7 @@ public class JobTriggerPoolHelper {
     /**
      * add trigger
      */
-    public void addTrigger(final int jobId, final TriggerTypeEnum triggerType, final int failRetryCount, final String executorShardingParam, final String executorParam) {
+    public void addTrigger(final Long jobId, final TriggerTypeEnum triggerType, final int failRetryCount, final String executorShardingParam, final String executorParam) {
 
         // choose thread pool
         ThreadPoolExecutor triggerPool_ = fastTriggerPool;
