+++ fast-data-switch-core/src/main/java/cn/zhumingwu/dataswitch/core/job/model/HandleCallbackParam.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
 public class HandleCallbackParam implements Serializable {
     private static final long serialVersionUID = 42L;
 
-    private int logId;
+    private Long logId;
     private long logDateTim;
 
     private Result<Object> executeResult;
